Extracted a test base class for SW360 clients.
Issue: eclipse#423

The class offers basic functionality to deal with the mock server and
process test JSON input and output.

SW360ComponentClientIT now uses the new base class.
